Started adding new server functions
