+++ src/common/Special.java
@@ -1,0 +1,5 @@
+package src.common;
+
+public enum Special {
+	USER, JOIN, CLIENT_END, SERVER_END, CRASH, USER_EXIT, SERVER_UP, FORWARD, SERVER, CLIENT, INFO
+}
