+++ src/common/State.java
@@ -1,0 +1,5 @@
+package src.common;
+
+public enum State {
+	START, CONNECTING, END, ERROR, RUNNING
+}
