--- src/common/States.java
@@ -1,5 +1,0 @@
-package src.common;
-
-public enum States {
-	START, END, SERVER_CLOSE, CLIENT_CLOSE, ERROR, RUNNING
-}
